pyproject.toml
[project]
name = "fastapillm"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "beanie>=1.29.0",
    "bs4>=0.0.2",
    "cachetools>=5.5.2",
    "deepeval>=2.7.5",
    "docx2txt>=0.8",
    "fastapi>=0.115.6",
    "firecrawl-py>=1.16.0",
    "gunicorn>=23.0.0",
    "joblib>=1.4.2",
    "jwt>=1.3.1",
    "langchain>=0.3.13",
    "langchain-community>=0.3.21",
    "langchain-core>=0.3.28",
    "langchain-google-genai>=2.1.4",
    "langchain-openai>=0.2.14",
    "langchain-pinecone>=0.2.0",
    "langgraph>=0.2.74",
    "motor>=3.6.0",
    "openapi-generator-cli>=7.12.0",
    "pinecone-client>=5.0.1",
    "pinecone-notebooks>=0.1.1",
    "pydantic-settings>=2.7.0",
    "pypdf>=5.1.0",
    "pypdf2>=3.0.1",
    "python-docx>=1.1.2",
    "python-multipart>=0.0.20",
    "redis>=5.2.1",
    "requests>=2.32.3",
    "scikit-learn>=1.6.1",
    "sentry-sdk>=2.19.2",
    "uvicorn>=0.34.0",
    "watchdog>=6.0.0",
]
